requirements.txt
python-can==3.3.4
can-isotp==1.7
udsoncan==1.13.1
pytest==6.2.4